% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gratis_app.R
\name{app_gratis}
\alias{app_gratis}
\title{Launch the gratis Shiny application}
\usage{
app_gratis()
}
\value{
Called for its side effect of launching a Shiny application. Returns
  the value from \code{\link[shiny]{runApp}()}.
}
\description{
Launch a local Shiny application for generating time series with
controllable features.
}
\examples{
\dontrun{
app_gratis()
}
}
