% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{arinf}
\alias{arinf}
\title{Compute pi coefficients from ARIMA model}
\usage{
arinf(object)
}
\arguments{
\item{object}{An object of class "Arima"}
}
\value{
A vector of AR coefficients
}
\description{
Compute pi coefficients from ARIMA model
}
\examples{
# Not Run
}
\author{
Rob J Hyndman
}
