% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Title}
\usage{
fitness(pars, features, seasonal, n = 120, freq = 12, target, nComp,
  selected.features)
}
\arguments{
\item{pars}{}

\item{features}{}

\item{seasonal}{}

\item{n}{}

\item{freq}{}

\item{target}{}

\item{nComp}{}

\item{selected.features}{}
}
\description{
Title
}
