% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ts.R
\name{generate_ts}
\alias{generate_ts}
\title{Generate random MAR time series}
\usage{
generate_ts(n.ts = 1, freq = 1, nComp = NULL, n = 120, output_format = "list")
}
\arguments{
\item{n.ts}{Number of time series to generate.}

\item{freq}{Seasonal period of the generated series.}

\item{nComp}{Number of mixture components. If \code{NULL}, this is sampled
randomly for each generated series.}

\item{n}{Length of each generated series.}

\item{output_format}{Output format: \code{"list"} for the historical list
format or \code{"tsibble"} for a list of tsibbles.}
}
\value{
When \code{output_format = "list"}, a list containing generated time
  series and the parameters used for each series. When
  \code{output_format = "tsibble"}, a list of tsibbles.
}
\description{
This function is deprecated. Use \code{\link{mar_model}()} with
\code{\link{generate.mar}()} or \code{\link{simulate.mar}()} instead.
}
\examples{
x <- generate_ts(n.ts = 2, freq = 12, nComp = 2, n = 120)
x$N1$pars
forecast::autoplot(x$N1$x)
}
\references{
Wong, CS & WK Li (2000).
}
\author{
Yanfei Kang and Feng Li
}
