% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gratis-package.R
\docType{package}
\name{gratis-package}
\alias{gratis}
\alias{gratis-package}
\title{gratis: Generate synthetic time series}
\description{
Generate synthetic univariate time series with diverse or controllable
characteristics. The package supports mixture autoregressive (MAR), ARIMA,
and ETS data-generating processes, and includes a genetic-algorithm workflow
for searching MAR parameters that produce user-specified time-series
features.
}
\details{
The main workflow is to construct a model with \code{\link{mar_model}()},
\code{\link{arima_model}()}, or \code{\link{ets_model}()}, then simulate one
path with \code{\link[stats]{simulate}()} or many paths with
\code{\link[generics]{generate}()}. To generate series with target feature
values, use \code{\link{simulate_target}()} or
\code{\link{generate_target}()}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ykang/gratis}
  \item Report bugs at \url{https://github.com/ykang/gratis/issues/}
}

}
\author{
\strong{Maintainer}: Feng Li \email{feng.li@gsm.pku.edu.cn} (\href{https://orcid.org/0000-0002-4248-9778}{ORCID})

Authors:
\itemize{
  \item Yanfei Kang \email{yanfeikang@buaa.edu.cn} (\href{https://orcid.org/0000-0001-8769-6650}{ORCID})
  \item Rob Hyndman \email{rob.hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

Other contributors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID}) [contributor]
  \item Bocong Zhao \email{bocongzhao@163.com} (\href{https://orcid.org/0000-0001-8434-9047}{ORCID}) [contributor]
}

}
\keyword{internal}
