% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gratis.R
\docType{package}
\name{gratis}
\alias{gratis}
\title{gratis: Generating Time Series with Diverse and Controllable Characteristics}
\description{
The gratis package generates synthetic time series data based on various
univariate time series models including MAR, ARIMA and ETS processes.
}
