% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGA.R
\name{myGA}
\alias{myGA}
\title{Title}
\usage{
myGA(type = c("binary", "real-valued", "permutation"), fitness, ..., n,
  min, max, nBits, population = gaControl(type)$population,
  selection = gaControl(type)$selection,
  crossover = gaControl(type)$crossover,
  mutation = gaControl(type)$mutation, popSize = 50,
  pcrossover = 0.8, pmutation = 0.1, elitism = base::max(1,
  round(popSize * 0.05)), updatePop = FALSE, postFitness = NULL,
  maxiter = 100, run = maxiter, maxFitness = Inf, names = NULL,
  suggestions = NULL, optim = FALSE, optimArgs = list(method =
  "L-BFGS-B", poptim = 0.05, pressel = 0.5, control = list(fnscale = -1,
  maxit = 100)), keepBest = FALSE, parallel = FALSE, monitor = if
  (interactive()) {     if (is.RStudio())          gaMonitor     else FALSE
  } else FALSE, seed = NULL)
}
\arguments{
\item{type}{}

\item{fitness}{}

\item{...}{}

\item{n}{}

\item{min}{}

\item{max}{}

\item{nBits}{}

\item{population}{}

\item{selection}{}

\item{crossover}{}

\item{mutation}{}

\item{popSize}{}

\item{pcrossover}{}

\item{pmutation}{}

\item{elitism}{}

\item{updatePop}{}

\item{postFitness}{}

\item{maxiter}{}

\item{run}{}

\item{maxFitness}{}

\item{names}{}

\item{suggestions}{}

\item{optim}{}

\item{optimArgs}{}

\item{keepBest}{}

\item{parallel}{}

\item{monitor}{}

\item{seed}{}
}
\description{
Title
}
