% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nsdiffs1}
\alias{nsdiffs1}
\title{Set the number of seasonal differences for yearly data to be -1.}
\usage{
nsdiffs1(x)
}
\arguments{
\item{x}{Univariate time series or numerical vector}
}
\value{
A numerical scalar value
}
\description{
Set the number of seasonal differences for yearly data to be -1.
}
\examples{
# Not Run
}
