% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm2.R
\name{rlnorm2}
\alias{rlnorm2}
\title{Log-normal distribution with alternative parametrization.}
\usage{
rlnorm2(n, mean, sd)
}
\arguments{
\item{n}{number of observations.}

\item{mean}{"vector" the mean value of the log-normal distribution.}

\item{sd}{"vector" the variance of the log-normal distribution.}
}
\value{
See the corresponding help for the usual log-normal functions.
}
\description{
Alternative parametrization of log normal distribution.
}
\details{
See help("rlnorm") for the details for the log-normal distribution.
}
\references{
Feng Li, Mattias Villani, and Robert Kohn. (2010). Flexible Modeling of
    Conditional Distributions using Smooth Mixtures of Asymmetric Student T Densities,
    Journal of Statistical Planning and Inference, 140(12), pp. 3638-3654
}
\author{
Feng Li, Central University of Finance and Economics .
}
