% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.mixture.data.R
\name{simulate_mixture_data}
\alias{simulate_mixture_data}
\title{Simulate time series from random parameter spaces of the mixture autoregressive (MAR) models.}
\usage{
simulate_mixture_data(n.ts = 1, freq = 1, nComp = NULL, n = 120)
}
\arguments{
\item{n.ts}{number of time series to be simulated.}

\item{freq}{seasonal period of the time series to be simulated.}

\item{nComp}{number of mixing components when simulating time series using MAR models.}

\item{n}{length of the simulated time series.}
}
\value{
a list of time series together with the SARIMA coefficients used in each mixing
component and the corresponding mixing weights.
}
\description{
Simulate time series from random parameter spaces of the mixture autoregressive (MAR) models.
}
\examples{
x <- simulate_mixture_data(n.ts = 2, freq = 12, nComp = 2, n = 120)
x$N1$pars
plot(x$N1$x)
}
\references{
Wong, CS & WK Li (2000).
}
\author{
Yanfei Kang and Feng Li
}
