% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ts.R
\name{simulate_msts}
\alias{simulate_msts}
\title{Simulate mutiple seasonal time series from random parameter spaces of the mixture autoregressive (MAR) models.}
\usage{
simulate_msts(seasonal.periods = c(7, 365), n = 800, nComp = NULL)
}
\arguments{
\item{seasonal.periods}{a vector of seasonal periods of the time series to be simulated.}

\item{n}{length of the simulated time series.}

\item{nComp}{number of mixing components when simulating time series using MAR models.}
}
\value{
a time series with multiple seasonal periods.
}
\description{
Simulate mutiple seasonal time series from random parameter spaces of the mixture autoregressive (MAR) models.
}
\examples{
x <- simulate_msts(seasonal.periods = c(7, 365), n = 800, nComp = 2)
plot(x)
}
