% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsgeneration.R
\docType{package}
\name{tsgeneration}
\alias{tsgeneration}
\alias{tsgeneration-package}
\title{Time Series Generation}
\description{
The tsgeneration package generates time series data based on MAR models.
}
