% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gratis_app.R
\name{run_gratis_app}
\alias{run_gratis_app}
\alias{app_gratis}
\title{Web Application to generate time series with controllable features.}
\usage{
run_gratis_app()

app_gratis()
}
\description{
Web Application to generate time series with controllable features.
}
\examples{
# Not Run
}
