% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gratis_app.R
\name{app_gratis}
\alias{app_gratis}
\title{Web Application to generate time series with controllable features.}
\usage{
app_gratis()
}
\description{
Web Application to generate time series with controllable features.
}
