% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Fitness function for time series generation.}
\usage{
fitness(pars, features, seasonal, n = 120, freq = 12, target, nComp,
  selected.features)
}
\arguments{
\item{pars}{Parameters}

\item{features}{Time series features.}

\item{seasonal}{Seasonal effects.}

\item{n}{Length of time series}

\item{freq}{Frequence of time series}

\item{target}{Target time series features}

\item{nComp}{No. of components used in mixture models.}

\item{selected.features}{Selected features.}
}
\value{
NA
}
\description{
Fitness function for time series generation.
}
\examples{
# Not Run
}
