% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ts.R, R/simulate_ts.R
\name{generate_ts}
\alias{generate_ts}
\title{Generating diverse time series}
\usage{
generate_ts(n.ts = 1, freq = 1, nComp = NULL, n = 120)

generate_ts(n.ts = 1, freq = 1, nComp = NULL, n = 120)
}
\arguments{
\item{n.ts}{number of time series to be simulated.}

\item{freq}{Frequence}

\item{nComp}{number of mixing components when simulating time series using MAR models.}

\item{n}{Length of time series}

\item{selected.features}{Selected features.}

\item{target}{Target features.}

\item{seasonal}{Seasonal effects.}

\item{ts.length}{Length}

\item{freq}{seasonal period of the time series to be simulated.}

\item{n}{length of the simulated time series.}
}
\value{
NA

a list of time series together with the SARIMA coefficients used in each mixing
component and the corresponding mixing weights.
}
\description{
Generate time series from random parameter spaces of the mixture autoregressive (MAR) models.
}
\examples{
# Not Run
x <- generate_ts(n.ts = 2, freq = 12, nComp = 2, n = 120)
x$N1$pars
plot(x$N1$x)
}
\references{
Wong, CS & WK Li (2000).
}
\author{
Yanfei Kang and Feng Li
}
