% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ts.R
\name{generate_ts}
\alias{generate_ts}
\title{Title}
\usage{
generate_ts(n, ts.length, freq, seasonal, features, selected.features,
  target)
}
\arguments{
\item{n}{}

\item{ts.length}{}

\item{freq}{}

\item{seasonal}{}

\item{selected.features}{}

\item{target}{}
}
\description{
Title
}
