% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ts_with_target.R
\name{generate_ts_with_target}
\alias{generate_ts_with_target}
\title{Generating time series with controllable features.}
\usage{
generate_ts_with_target(n, ts.length, freq, seasonal, features,
  selected.features, target)
}
\arguments{
\item{n}{number of time series to be generated.}

\item{ts.length}{length of the time series to be generated.}

\item{freq}{frequency of the time series to be generated.}

\item{seasonal}{0 for non-seasonal data, 1 for single-seasonal data, and 2 for multiple seasonal data.}

\item{features}{a vector of function names.}

\item{selected.features}{selected features to be controlled.}

\item{target}{target feature values.}
}
\value{
A time-series object of class "ts" or "msts".
}
\description{
Generating time series with controllable features.
}
\examples{
library(tsfeatures)
x <- generate_ts_with_target(n = 1, ts.length = 60, freq = 1, seasonal = 0, features = c('entropy', 'stl_features'), selected.features = c('entropy', 'trend'), c(0.6, 0.9))
forecast::autoplot(x)
}
\author{
Yanfei Kang
}
