% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gratis-package.R
\docType{package}
\name{gratis-package}
\alias{gratis-package}
\alias{_PACKAGE}
\title{gratis: GeneRAting TIme Series with diverse and controllable characteristics}
\description{
The gratis package generates synthetic time series data based on various
univariate time series models including MAR, ARIMA and ETS processes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ykang/gratis}
  \item Report bugs at \url{https://github.com/ykang/gratis/issues/}
}

}
\author{
\strong{Maintainer}: Feng Li \email{feng.li@cufe.edu.cn} (\href{https://orcid.org/0000-0002-4248-9778}{ORCID})

Authors:
\itemize{
  \item Yanfei Kang \email{yanfeikang@buaa.edu.cn} (\href{https://orcid.org/0000-0001-8769-6650}{ORCID})
  \item Rob Hyndman \email{rob.hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

Other contributors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID}) [contributor]
  \item Bocong Zhao \email{bocongzhao@163.com} (\href{https://orcid.org/0000-0001-8434-9047}{ORCID}) [contributor]
}

}
\keyword{internal}
