% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_target.R
\name{simulate_target}
\alias{simulate_target}
\title{Generating time series with controllable features.}
\usage{
simulate_target(
  length,
  seasonal_periods = 1,
  feature_function,
  target,
  parallel = FALSE
)
}
\arguments{
\item{length}{length of the time series to be generated.}

\item{seasonal_periods}{Either a scalar or a numeric vector of length k containing
the number of seasonal periods for each component.}

\item{feature_function}{a function that returns a vector of features from a time series}

\item{target}{target feature values of the same length as that returned by \code{feature_function()}}

\item{parallel}{An optional argument which allows to specify if the Genetic Algorithm
should be run sequentially or in parallel.}
}
\value{
A time-series object of class "ts" or "msts".
}
\description{
Generating time series with controllable features.
}
\examples{
\dontrun{
my_features <- function(y) {
  c(tsfeatures::entropy(y), acf = acf(y, plot = FALSE)$acf[2:3, 1, 1])
}
y <- simulate_target(
  length = 60, feature_function = my_features,
  target = c(0.5, 0.9, 0.8)
)
my_features(y)
plot(y)}
}
\author{
Yanfei Kang and Rob J Hyndman
}
